requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
redis==5.0.3
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.1.1
httpx==0.27.0
requests==2.31.0

redislite==6.2.912183
